+++ app/proguard-rules.pro
@@ -160,3 +160,25 @@
 # 若启用模糊效果，请增加如下配置：
 -dontwarn androidx.renderscript.**
 -keep public class androidx.renderscript.** { *; }
+
+
+-keepattributes SourceFile,LineNumberTable
+-keep class com.inmobi.** { *; }
+-keep public class com.google.android.gms.**
+-dontwarn com.google.android.gms.**
+-dontwarn com.squareup.picasso.**
+-keep class com.google.android.gms.ads.identifier.AdvertisingIdClient{
+     public *;
+}
+-keep class com.google.android.gms.ads.identifier.AdvertisingIdClient$Info{
+     public *;
+}
+# skip the Picasso library classes
+-keep class com.squareup.picasso.** {*;}
+-dontwarn com.squareup.okhttp.**
+# skip Moat classes
+-keep class com.moat.** {*;}
+-dontwarn com.moat.**
+# skip IAB classes
+-keep class com.iab.** {*;}
+-dontwarn com.iab.**
