+++ app/release/output-metadata.json
@@ -11,8 +11,8 @@
       "type": "SINGLE",
       "filters": [],
       "attributes": [],
-      "versionCode": 5,
-      "versionName": "1.2.2 beta",
+      "versionCode": 8,
+      "versionName": "1.2.5 beta",
       "outputFile": "app-release.apk"
     }
   ],
