+++ app/src/main/java/com/lamesa/socialdown/ui/viewmodel/MainViewModelFactory.kt
@@ -1,0 +1,15 @@
+package com.lamesa.socialdown.ui.viewmodel
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+
+class MainViewModelFactory : ViewModelProvider.Factory {
+
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(MainViewModel::class.java)) {
+            return MainViewModel() as T
+        }
+        throw IllegalArgumentException("Unknown ViewModel Class")
+    }
+
+}
