--- app/src/main/java/com/lamesa/socialdown/viewmodel/MainViewModelFactory.kt
@@ -1,15 +1,0 @@
-package com.lamesa.socialdown.viewmodel
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-
-class MainViewModelFactory : ViewModelProvider.Factory {
-
-    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(MainViewModel::class.java)) {
-            return MainViewModel() as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel Class")
-    }
-
-}
